requirements.txt
beautifulsoup4==4.13.3
bertopic==0.16.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.17.0
Flask==3.1.0
fsspec==2025.2.0
gunicorn==23.0.0
hdbscan==0.8.40
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.27.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
praw==7.8.1
prawcore==2.4.0
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sympy==1.13.1
textblob==0.19.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
